requirements.txt
azure-core==1.22.1
azure-data-tables==12.2.0
azure-functions==1.8.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
idna==3.3
isodate==0.6.1
msrest==0.6.21
oauthlib==3.2.0
pycparser==2.21
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.8
